[Faucet] Remove service layer
