+++ faucet/src/requests.rs
@@ -1,0 +1,23 @@
+// Copyright (c) 2022, Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+use sui_types::base_types::SuiAddress;
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub enum FaucetRequest {
+    FixedAmountRequest(FixedAmountRequest),
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct FixedAmountRequest {
+    pub recipient: SuiAddress,
+}
+
+impl FaucetRequest {
+    pub fn new_fixed_amount_request(recipient: impl Into<SuiAddress>) -> Self {
+        Self::FixedAmountRequest(FixedAmountRequest {
+            recipient: recipient.into(),
+        })
+    }
+}
